user profile

+++ utils/otp.util.js
@@ -2,13 +2,50 @@ import nodemailer from "nodemailer";
 import Otp from "../models/otp.model.js";
 import crypto from "crypto";
 
+export const otpEmailTemplates = {
+  signup: {
+    subject: "Verify your LuxeTime account",
+    getHtml: (otp) => `
+      <h2>Welcome to LuxeTime </h2>
+      <p>Your verification code is:</p>
+      <h1 style="letter-spacing:4px">${otp}</h1>
+      <p>This OTP is valid for <b>5 minutes</b>.</p>`
+  },
 
-const generateOtp=()=>{
-    return Math.floor(100000+Math.random()*90000).toString();
+  "forgot-password": {
+    subject: "Reset your LuxeTime password",
+    getHtml: (otp) => `
+      <h2>Password Reset Request</h2>
+      <p>Use the OTP below to reset your password:</p>
+      <h1 style="letter-spacing:4px">${otp}</h1>
+      <p>This OTP expires in <b>5 minutes</b>.</p>
+      <p>If you didn’t request this, ignore this email.</p>`
+  },
+  
+    "change-email": {
+    subject: "Reset your LuxeTime email",
+    getHtml: (otp) => `
+      <h2>Email Reset Request</h2>
+      <p>Use the OTP below to reset your email:</p>
+      <h1 style="letter-spacing:4px">${otp}</h1>
+      <p>This OTP expires in <b>5 minutes</b>.</p>
+      <p>If you didn’t request this, ignore this email.</p>`
+    },
 };
 
-const sendOtpMail = async (email,otp)=>{
+
+
+const generateOtp=()=>{
+    return Math.floor(100000+Math.random()*900000).toString();
+};
+
+const sendOtpMail = async (email,otp,purpose)=>{
     try{
+
+        // console.log(purpose);
+        // console.log(email);
+
+        const template=otpEmailTemplates[purpose];
     const transporter =nodemailer.createTransport({
         service:"gmail",
         auth:{
@@ -20,8 +57,8 @@ const sendOtpMail = async (email,otp)=>{
     await transporter.sendMail({
         from:`LuxeTime <${process.env.NODEMAILER_EMAIL}>`,
         to:email,
-        subject:"Verify your account",
-        html:`<h2> Your OTP is ${otp} </h2> <p> valid for 5 minutes </p>`,
+        subject:template.subject,
+        html:template.getHtml(otp)
     });
 
     console.log("otp sented");
@@ -33,7 +70,10 @@ const sendOtpMail = async (email,otp)=>{
 
 
 
-export const createAndSendOtp=async(user)=>{
+export const createAndSendOtp=async(user,purpose)=>{
+    if(!user?.email){
+        throw new Error("OTP email missing");
+    }
 
     await Otp.deleteMany({ userId: user._id });
     
@@ -47,5 +87,5 @@ export const createAndSendOtp=async(user)=>{
     });
 
 
-    await sendOtpMail(user.email,otpCode);
+    await sendOtpMail(user.email,otpCode,purpose);
 }
